Renaming files and cleaning up the externallogincallback endpoint
