+++ src/Template.Infrastructure/Identity/Services/IdentityService.cs
@@ -262,7 +262,7 @@ public sealed class IdentityService(
             var info = new UserLoginInfo(provider, userId, provider);
 
             await _userManager.AddLoginAsync(user, info).ConfigureAwait(false);
-            // await _signInManager.SignInAsync(user, isPersistent: false).ConfigureAwait(false);
+            await _signInManager.SignInAsync(user, isPersistent: false).ConfigureAwait(false);
 
             return Result<object>.Success();
         }
